First draft of a unified ssl_simulator protocol
This protocol is supposed to be used to control a robot,
to configure a simulator for use in another application,
and to manipulate the state of the simulation for testing purposes.

The protocol is supposed to be implemented by as many simulators as
possible, giving teams more options to change their simulator or write
their own instead of "locking in" to a specific protocol for a specific
implementation.

However, we think the configuration of the simulation itself
cannot be done purely generic. Instead, each simulator vendor is asked
to publish their own set of specific configuration, mostly for realism
features, and we will add them here as an alternative to the current
ERForce configuration command.